Created default atomix client
Also created the operation types that can be executed. All will be in
parts, requests will be sent by the SharderPrimitiveClient, then will
compose with AtomixClient that will choose to which node this request
will be sent. The socket listener then will sent the request to the
AtomixInvoker, then with the AtomixWrapper the request finally will be
executed.